games_services/requirements.txt
Django==2.1.4
djangorestframework==3.9.0
httpie==0.9.9
psycopg2==2.7.5